Island: Move JSONFileUserRepository to the repository package
