+++ monkey/monkey_island/cc/repository/__init__.py
@@ -17,4 +17,5 @@ from .file_repository_logging_decorator import FileRepositoryLoggingDecorator
 from .agent_binary_repository import AgentBinaryRepository
 from .file_agent_configuration_repository import FileAgentConfigurationRepository
 from .file_simulation_repository import FileSimulationRepository
+from .json_file_user_repository import JSONFileUserRepository
 from .mongo_credentials_repository import MongoCredentialsRepository
