+++ monkey/monkey_island/cc/repository/json_file_user_repository.py
